pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code_reviewer"
version = "0.1.0"
description = "Code Reviewer with LLM integrations"
authors = [
    { name="Soumik Deb Niloy", email="you@example.com" }
]
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "pydantic",
    "motor",
    "starlette",
    "requests",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["LLMIntegrations*", "backend*"]
